# static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "API for utilizing case study application functionality",
        "version": "1.0",
        "title": "Bank Case Study API",
        "license": {
            "name": "CC0 1.0 Universal",
            "url": "https://creativecommons.org/publicdomain/zero/1.0/"
        }
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "components": {
        "schemas": {
            "RetrieveToken": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
            "TokenResult": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string"
                    },
                    "success": {
                        "type": "boolean"
                    }
                }
            },
            "VerifyToken": {
                "type" : "string",
                "properties" : {
                    "apikey" : {
                        "type" : "string"
                    }
                }
            },
...
            "Success": {
                "type" : "object",
                "properties" : {
                    "message" : {
                        "type" : "string"
                    }
                }
            },
            "AddAccount": {
                "type" : "object",
                "properties" : {
                    "account_type" : {
                        "type" : "string"
                    },
                    "password" : {
                        "type" : "string"
                    },
                    "balance" : {
                        "type" : "integer"
                    }
                }
            },
...
            "AddAccountResult": {
                "type" : "object",
                "properties" : {
                    "message" : {
                        "type" : "string"
                    },
                    "account" : {
                        "type" : "string"
                    }
                }
            },
            "ViewAccountResult": {
                "type" : "object",
                "properties" : {
                    "message" : {
                        "type" : "string"
                    },
                    "accounts" : {
                        "type" : "string"
                    }
                }
            },
...
        }
    },
    "paths": {
        "/api/retrievetoken": {
            "post": {
                "tags": [
                    "Retrieve token API"
                ],
                "summary": "Retrieve token",
                "requestBody": {
                    "description": "Enter credentials to retrieve token",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RetrieveToken"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema" : {
                            "$ref": "#/components/schemas/TokenResult"
                        }
                    },
                    "400": {
                        "description": "BAD REQUEST",
                        "schema" : {
                            "$ref": "#/components/schemas/TokenResult"
                        }
                    },
                    "500": {
                        "description": "SERVER ERROR",
                        "schema" : {
                            "$ref": "#/components/schemas/TokenResult"
                        }
                    }
                }
            }
        },
        "/api/verifytoken": {
            "get": {
                "tags": [
                    "Verify token API"
                ],
                "summary": "Verify token",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "header",
                        "description": "Authentication token",
                        "required": true,
                        "type": "string",
                        "schema" : {
                            "$ref": "#/components/schemas/VerifyToken"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema" : {
                            "$ref": "#/components/schemas/Success"
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED",
                        "schema" : {
                            "$ref": "#/components/schemas/Success"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema" : {
                            "$ref": "#/components/schemas/Success"
                        }
                    },
                    "500": {
                        "description": "SERVER ERROR",
                        "schema" : {
                            "$ref": "#/components/schemas/Success"
                        }
                    }
                }
            }
        },
        "/api/addaccount": {
            "post": {
                "tags": [
                    "Add account API"
                ],
                "summary": "Add account",
                "requestBody": {
                    "description": "Enter account details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddAccount"
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "header",
                        "description": "Authentication token",
                        "required": true,
                        "type": "string",
                        "schema" : {
                            "$ref": "#/components/schemas/VerifyToken"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema" : {
                            "$ref": "#/components/schemas/AddAccountResult"
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED",
                        "schema" : {
                            "$ref": "#/components/schemas/Success"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema" : {
                            "$ref": "#/components/schemas/AddAccountResult"
                        }
                    },
                    "500": {
                        "description": "SERVER ERROR",
                        "schema" : {
                            "$ref": "#/components/schemas/AddAccountResult"
                        }
                    }
                }
            }
        },
        "/api/viewaccount": {
            "get": {
                "tags": [
                    "View accounts API"
                ],
                "summary": "View accounts",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "header",
                        "description": "Authentication token",
                        "required": true,
                        "type": "string",
                        "schema" : {
                            "$ref": "#/components/schemas/VerifyToken"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema" : {
                            "$ref": "#/components/schemas/ViewAccountResult"
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED",
                        "schema" : {
                            "$ref": "#/components/schemas/Success"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema" : {
                            "$ref": "#/components/schemas/ViewAccountResult"
                        }
                    },
                    "500": {
                        "description": "SERVER ERROR",
                        "schema" : {
                            "$ref": "#/components/schemas/ViewAccountResult"
                        }
                    }
                }
            }
        },
        "/api/viewtransactionhistory": {
            "get": {
                "tags": [
                    "View Transaction History API"
                ],
                "summary": "View transaction history",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "header",
                        "description": "Authentication token",
                        "required": true,
                        "type": "string",
                        "schema" : {
                            "$ref": "#/components/schemas/VerifyToken"
                        }
                    },
                    {
                        "name": "account_no",
                        "in": "header",
                        "description": "Account number",
                        "required": true,
                        "type": "integer",
                        "schema" : {
                            "$ref": "#/components/schemas/AccountNumber"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema" : {
                            "$ref": "#/components/schemas/ViewHistoryResult"
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED",
                        "schema" : {
                            "$ref": "#/components/schemas/Success"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema" : {
                            "$ref": "#/components/schemas/ViewHistoryResult"
                        }
                    },
                    "500": {
                        "description": "SERVER ERROR",
                        "schema" : {
                            "$ref": "#/components/schemas/ViewHistoryResult"
                        }
                    }
                }
            }
        },
        "/api/maketransaction": {
            "post": {
                "tags": [
                    "Make transaction API"
                ],
                "summary": "Make transaction",
                "requestBody": {
                    "description": "Enter transaction details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Transaction"
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "header",
                        "description": "Authentication token",
                        "required": true,
                        "type": "string",
                        "schema" : {
                            "$ref": "#/components/schemas/VerifyToken"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema" : {
                            "$ref": "#/components/schemas/Success"
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED",
                        "schema" : {
                            "$ref": "#/components/schemas/Success"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema" : {
                            "$ref": "#/components/schemas/Success"
                        }
                    },
                    "500": {
                        "description": "SERVER ERROR",
                        "schema" : {
                            "$ref": "#/components/schemas/Success"
                        }
                    }
                }
            }
        },
...
    }
}